refactor(solver): split evaluation api wrappers
