+++ crates/tsz-solver/src/evaluation/evaluate.rs
@@ -11,6 +11,7 @@
 //! - Handles deferred evaluation when type parameters are unknown
 //! - Supports distributivity for naked type parameters in unions
 
+mod api;
 pub(in crate::evaluation) mod application_types;
 mod array_methods;
 
@@ -30,11 +31,15 @@ use crate::relations::subtype::{NoopResolver, TypeResolver};
 #[cfg(test)]
 use crate::types::*;
 use crate::types::{
-    ConditionalType, ConditionalTypeId, MappedType, MappedTypeId, StringIntrinsicKind,
-    TemplateLiteralId, TemplateSpan, TupleElement, TupleListId, TypeApplicationId, TypeData,
-    TypeId, TypeListId, TypeParamInfo,
+    ConditionalTypeId, MappedTypeId, StringIntrinsicKind, TemplateLiteralId, TemplateSpan,
+    TupleElement, TupleListId, TypeApplicationId, TypeData, TypeId, TypeListId, TypeParamInfo,
 };
 use crate::visitors::visitor_predicates::contains_type_matching;
+pub use api::{
+    evaluate_conditional, evaluate_index_access, evaluate_index_access_with_options,
+    evaluate_keyof, evaluate_mapped, evaluate_type, evaluate_type_with_request,
+    evaluate_type_with_resolver,
+};
 use application_types::{ApplicationEvalContext, ApplicationEvalOutcome, HomomorphicMappedArg};
 pub(crate) use array_methods::{
     ARRAY_METHODS_RETURN_ANY, ARRAY_METHODS_RETURN_BOOLEAN, ARRAY_METHODS_RETURN_NUMBER,
@@ -45,15 +50,6 @@ use tsz_common::interner::Atom;
 
 mod closed_eval;
 mod support;
-
-/// Controls which subtype direction makes a member redundant when simplifying
-/// a union or intersection.
-enum SubtypeDirection {
-    /// member[i] <: member[j] → member[i] is redundant (union semantics).
-    SourceSubsumedByOther,
-    /// member[j] <: member[i] → member[i] is redundant (intersection semantics).
-    OtherSubsumedBySource,
-}
 
 /// Type evaluator for meta-types.
 ///
@@ -1988,70 +1984,6 @@ impl<'a, R: TypeResolver> TypeEvaluator<'a, R> {
     }
 
     // Additional evaluator support methods live in the nested support module.
-}
-
-/// Convenience function for evaluating conditional types
-pub fn evaluate_conditional(interner: &dyn TypeDatabase, cond: &ConditionalType) -> TypeId {
-    let mut evaluator = TypeEvaluator::new(interner);
-    evaluator.evaluate_conditional(cond)
-}
-
-/// Convenience function for evaluating index access types
-pub fn evaluate_index_access(
-    interner: &dyn TypeDatabase,
-    object_type: TypeId,
-    index_type: TypeId,
-) -> TypeId {
-    let mut evaluator = TypeEvaluator::new(interner);
-    evaluator.evaluate_index_access(object_type, index_type)
-}
-
-/// Convenience function for evaluating index access types with options.
-pub fn evaluate_index_access_with_options(
-    interner: &dyn TypeDatabase,
-    object_type: TypeId,
-    index_type: TypeId,
-    no_unchecked_indexed_access: bool,
-) -> TypeId {
-    let mut evaluator = TypeEvaluator::new(interner);
-    evaluator.set_no_unchecked_indexed_access(no_unchecked_indexed_access);
-    evaluator.evaluate_index_access(object_type, index_type)
-}
-
-/// Convenience function for full type evaluation
-pub fn evaluate_type(interner: &dyn TypeDatabase, type_id: TypeId) -> TypeId {
-    evaluate_type_with_request(interner, EvaluationRequest::new(type_id))
-}
-
-/// Convenience function for full type evaluation with an explicit resolver.
-pub fn evaluate_type_with_resolver(
-    interner: &dyn TypeDatabase,
-    resolver: &impl TypeResolver,
-    type_id: TypeId,
-) -> TypeId {
-    let mut evaluator = TypeEvaluator::with_resolver(interner, resolver);
-    evaluator.evaluate(type_id)
-}
-
-/// Convenience function for full type evaluation with explicit request options.
-pub fn evaluate_type_with_request(
-    interner: &dyn TypeDatabase,
-    request: EvaluationRequest,
-) -> TypeId {
-    let mut evaluator = TypeEvaluator::new(interner);
-    evaluator.evaluate_request_result(request).into_type_id()
-}
-
-/// Convenience function for evaluating mapped types
-pub fn evaluate_mapped(interner: &dyn TypeDatabase, mapped: &MappedType) -> TypeId {
-    let mut evaluator = TypeEvaluator::new(interner);
-    evaluator.evaluate_mapped(mapped)
-}
-
-/// Convenience function for evaluating keyof types
-pub fn evaluate_keyof(interner: &dyn TypeDatabase, operand: TypeId) -> TypeId {
-    let mut evaluator = TypeEvaluator::new(interner);
-    evaluator.evaluate_keyof(operand)
 }
 
 // Re-enabled evaluate tests - verifying API compatibility
